display: carve webrtc ice.rs (pure move)
